% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipumsexamples.r
\docType{package}
\name{ipumsexamples-package}
\alias{ipumsexamples}
\alias{ipumsexamples-package}
\title{Example data for the ipumsr package}
\description{
ipumsexamples is only used to bring a few data extrats from the ipums.org
website for use in examples and tests for the ipumsr package. These
files are too large to be included on CRAN and so have been put in
a separate package.
}
\author{
\strong{Maintainer}: Minnesota Population Center \email{gfellis@umn.edu}

}
